Fix: calendar 권한 없을 시 오류 수정 등

+++ backend/src/main/java/com/dumpit/repository/ActivityLogRepository.java
@@ -13,5 +13,5 @@ public interface ActivityLogRepository extends JpaRepository<ActivityLog, UUID> 
 
     @Modifying
     @Query("DELETE FROM ActivityLog l WHERE l.user.withdrawnAt < :cutoff")
-    long deleteWithdrawnUserLogsBefore(@Param("cutoff") LocalDateTime cutoff);
+    int deleteWithdrawnUserLogsBefore(@Param("cutoff") LocalDateTime cutoff);
 }
